fix(metrics): review 01-25
